feat: config: make clap dependency optional
Move `clap` and `byte-unit` behind a new `clap` feature flag (kept in
`default = [...]` for backward compatibility). Library users who never
parse `Config` from CLI args can opt out and drop both dependencies
entirely.

The previous `Default for Config` impl ran clap's parser at runtime
on an empty arg list — `<Self as Parser>::parse_from(Vec::<&'static
str>::new())` — to recover compile-time-known defaults. That coupled
`Default` to clap and incurred a parse-and-validate cycle on every
`Config::default()`. It is now direct field assignment, sourced from
a new internal `Defaults` struct + `DEFAULTS` const that also feeds
clap's `default_value_t` attributes — single source of truth for both
code paths.

`Config::build`, `parse_bytes_with_unit`, and `parse_snapshot_policy`
are gated behind the `clap` feature. Clap-only tests moved into
`config_clap_test.rs` so the feature gate applies once at module
scope instead of repeating on each `#[test]`.

- Fix: #1742
