+++ openraft/src/docs/feature_flags/feature-flags-toc.md
@@ -1,6 +1,7 @@
 - [feature-flag `adapt-network-v1` (removed)](#feature-flag-adapt-network-v1-removed)
 - [feature-flag `bench`](#feature-flag-bench)
 - [feature-flag `bt`](#feature-flag-bt)
+- [feature-flag `clap`](#feature-flag-clap)
 - [feature-flag `compat`](#feature-flag-compat)
 - [feature-flag `loosen-follower-log-revert` (removed)](#feature-flag-loosen-follower-log-revert-removed)
 - [feature-flag `metrics-logids`](#feature-flag-metrics-logids)
